removing files and such
