+++ src/components/Bpm.vue
@@ -7,12 +7,10 @@
 
 <script lang="ts">
 import { Component, Prop, Mixins } from 'vue-property-decorator';
-import { DragElement } from '@/draggable';
 import { createComponent } from '@vue/composition-api';
 
 export default createComponent({
   name: 'Bpm',
-  components: { DragElement },
   props: {
     value: { type: Number, required: true },
   },
