--- src/draggable/index.ts
@@ -1,12 +1,0 @@
-import DragElement from '@/draggable/DragElement.vue';
-
-export default {
-  // tslint:disable-next-line:no-shadowed-variable
-  install(Vue: any) {
-    Vue.component('DragElement', DragElement);
-  },
-};
-
-export {
-  DragElement,
-};
