--- src/knobs/bus.ts
@@ -1,3 +1,0 @@
-import Vue from 'vue';
-
-export default new Vue();
